+++ packages/akashic-cli-commons/src/NodeModules.ts
@@ -59,7 +59,14 @@ export module NodeModules {
 		return moduleMainScripts;
 	}
 
-	export function listScriptFiles(basepath: string, modules: string|string[], logger: Logger): Promise<string[]> {
+	// TODO: node_modules/ 以下以外でも利用するメソッドのため、NodeModules ではなく別の適切な場所に移動する
+	// checkAllModules は実験的なものです。 akashic-cli の外部から利用しないでください
+	export function listScriptFiles(
+		basepath: string,
+		modules: string|string[],
+		logger: Logger,
+		checkAllModules: boolean = false
+	): Promise<string[]> {
 		if (modules.length === 0) return Promise.resolve([]);
 		var moduleNames = (typeof modules === "string") ? [modules] : modules;
 
@@ -87,7 +94,7 @@ export module NodeModules {
 			var filePaths: string[] = [];
 			b.on("dep", (row: any) => {
 				var filePath = Util.makeUnixPath(path.relative(basepath, row.file));
-				if (!(/^(?:\.\/)?node_modules/.test(filePath))) {
+				if (!checkAllModules && !(/^(?:\.\/)?node_modules/.test(filePath))) {
 					return;
 				}
 				if (/^\.\.\//.test(filePath)) {