+++ packages/akashic-cli-export/spec/fixtures/simple_game_with_operation_plugins/script/main.js
@@ -1,0 +1,10 @@
+var foo = require("./foo");
+var hoge = require("@hoge/testmodule/lib");
+
+function main() {
+	return {
+		x: foo()
+	};
+}
+
+module.exports = main;